Updating single champion functions to only refresh the handler for that champion when using SH 4.0.1 or later.
ScriptHub version 4.0.1 has HeroHandlers rewritten to take not need to reload all handlers when passed their effect key string to the Refresh() method.

This fix allows LevelUp to take advantage of the new performance improvements.

Note: When no effect key is passed, the Refresh()) method functions the same as before, so not using it does not break the addon.